Баланс и списания

+++ migrations/002_create_orders.sql
@@ -2,6 +2,6 @@ CREATE TABLE IF NOT EXISTS orders (
     number TEXT PRIMARY KEY,
     user_id BIGINT NOT NULL REFERENCES users(id),
     status TEXT NOT NULL DEFAULT 'NEW',
-    accrual NUMERIC(12,2),
+    accrual BIGINT,
     uploaded_at TIMESTAMPTZ NOT NULL DEFAULT now()
 );
