+++ migrations/003_create_withdrawals.sql
@@ -2,6 +2,6 @@ CREATE TABLE IF NOT EXISTS withdrawals (
     id BIGSERIAL PRIMARY KEY,
     user_id BIGINT NOT NULL REFERENCES users(id),
     order_number TEXT NOT NULL,
-    sum NUMERIC(12,2) NOT NULL,
+    sum BIGINT NOT NULL,
     processed_at TIMESTAMPTZ NOT NULL DEFAULT now()
 );
